requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.4.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.5
django-celery-beat==2.2.1
django-heroku==0.3.1
django-redis==5.0.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
kombu==5.1.0
pipenv==2021.5.29
prompt-toolkit==3.0.19
psycopg2==2.9.1
PyJWT==2.1.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
vine==5.0.0
virtualenv==20.4.7
virtualenv-clone==0.5.4
wcwidth==0.2.5
whitenoise==5.2.0
